Deprecating the usage of Telnet protocol and using TCPSocket instead
Telnet protocol was used in order to forward the toolsHCK.ps1 script
functionality by opening an interactive shell, but since Microsoft
declared Telnet functionality deprecated on Windows Server 2016 and
higher, we also need to deprecate the usage of Telnet protocol.

This commit holds the complete patch to convert the usage of Telnet
to using TCPSocket and toolsHCK.ps1 in server mode without sacrificing
any of the functionalities that RToolsHCK provides.

NOTE: toolsHCK.ps1 should be up to date!

Signed-off-by: Bishara AbuHattoum <bishara@daynix.com>